# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College LMS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /faculty:
    get:
      operationId: listFaculty
      tags: [faculty]
      summary: List all faculty
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of faculty
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faculty"
    post:
      operationId: createFaculty
      tags: [faculty]
      summary: Create faculty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacultyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"
  /faculty/{id}:
    get:
      operationId: getFaculty
      tags: [faculty]
      summary: Get faculty by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Faculty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"
    patch:
      operationId: updateFaculty
      tags: [faculty]
      summary: Update faculty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacultyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"
    delete:
      operationId: deleteFaculty
      tags: [faculty]
      summary: Delete faculty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # DEPARTMENTS
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List study materials
      parameters:
        - name: subjectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: semester
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Upload study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    delete:
      operationId: deleteMaterial
      tags: [materials]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # NOTICES
  /notices:
    get:
      operationId: listNotices
      tags: [notices]
      summary: List notices
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: pinned
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Notices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"
    post:
      operationId: createNotice
      tags: [notices]
      summary: Create notice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
  /notices/{id}:
    patch:
      operationId: updateNotice
      tags: [notices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    delete:
      operationId: deleteNotice
      tags: [notices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # EVENTS
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave applications
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Leave applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Leave"
    post:
      operationId: applyLeave
      tags: [leaves]
      summary: Apply for leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"
  /leaves/{id}/status:
    patch:
      operationId: updateLeaveStatus
      tags: [leaves]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leave"

  # DASHBOARD ANALYTICS
...
components:
  schemas:
...
    Faculty:
      type: object
      required: [id, name, email, employeeId, departmentId]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        employeeId:
          type: string
        phone:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        departmentId:
          type: integer
        departmentName:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        joinedAt:
          type: string
    FacultyInput:
      type: object
      required: [name, email, employeeId, departmentId, password]
      properties:
        name:
          type: string
        email:
          type: string
        employeeId:
          type: string
        phone:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        departmentId:
          type: integer
        password:
          type: string
    FacultyUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
...
    Material:
      type: object
      required: [id, title, subjectId, type, uploadedBy]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        type:
          type: string
        fileUrl:
          type: ["string", "null"]
        uploadedBy:
          type: integer
        uploaderName:
          type: ["string", "null"]
        createdAt:
          type: string
        semester:
          type: ["integer", "null"]
    MaterialInput:
      type: object
      required: [title, subjectId, type, uploadedBy]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        subjectId:
          type: integer
        type:
          type: string
        fileUrl:
          type: ["string", "null"]
        uploadedBy:
          type: integer
        semester:
          type: ["integer", "null"]
    Notice:
      type: object
      required: [id, title, content, createdBy, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        semester:
          type: ["integer", "null"]
        createdBy:
          type: integer
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
    NoticeInput:
      type: object
      required: [title, content, createdBy]
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        departmentId:
          type: ["integer", "null"]
        semester:
          type: ["integer", "null"]
        createdBy:
          type: integer
    NoticeUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        isPinned:
          type: ["boolean", "null"]
...
    Leave:
      type: object
      required: [id, studentId, fromDate, toDate, reason, status]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        fromDate:
          type: string
        toDate:
          type: string
        reason:
          type: string
        status:
          type: string
        reviewedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
    LeaveInput:
      type: object
      required: [studentId, fromDate, toDate, reason]
      properties:
        studentId:
          type: integer
        fromDate:
          type: string
        toDate:
          type: string
        reason:
          type: string
    LeaveStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        reviewedBy:
          type: ["integer", "null"]
...